requirements.txt
# streamlit # Front-end
# google-genai # LLM
# python-dotenv # To load env variables

streamlit>=1.24.1
python-dotenv>=1.0.0
PyPDF2>=3.0.1
langchain>=0.3.0
langchain-text-splitters>=0.2.0
langchain-community>=0.3.0
langchain-google-genai>=4.0.0
faiss-cpu>=1.8.0